Deprecate insufficient Display#setRescalingAtRuntime() method
The method Display#setRescalingAtRuntime() allows to activate the
monitor-specific scaling mechanism on Windows. This must, however,
actually be activated before the Display is instantiated, like it is
done during the Display constructor execution if the system property for
global activation of monitor-specific scaling is specified. In
consequence, calling #setRescalingAtRuntime() does not take the full
intended effect.

This change deprecates the method so that the system property needs to
be set for proper initialization of the monitor-specific scaling
functionality. It also adapts the according tests to the deprecation of
the API and alternatives to be used.