Use AutoScaleMethod "Smooth" by default
Currently, AutoScaleMethod "Nearest" is used by default, leading to
jagged image scaling results. The method "Smooth" is only used on GTK
with a fractional scaling value not being a multiple of 100.
The reason for not using "Smooth" scaling on all platform was a bug
regarding icon transparency that has long been resolved.

With this change, AutoScaleMethod "Smooth" is used by default and
"Nearest" can be activated via system property.