Embed Skija libraries in Tycho build
This change adds the Skija libraries to the configuration for the Tycho
build of the x64 architecture fragments for Linux and Windows and the
ARM architecture fragment for Windows (since we only have added JARs for
these architectures so far). It also limits the Tycho build to the
Windows fragment for now, as the others face compilation issues. In
consequence of this change, the Tycho build without tests will work. The
change is limited to executing the build and will not allow to use the
generated JARs (as they will not have the Skija JARs embedded so far).