Make TextLayout use consistent font size
The TextLayout uses some heuristics to calculate the proper font size
for rendering StyledText with Skija. One of the calculations that is
just based on the original SWT font data can be replaced by the existing
font size calculation in SkijaGC. Among others, this corrects faulty
font sizes in Trees.
This change also adapts the retrieval of according font data throughout
the TextLayout in general, making the result more consistent.