Unify and improve image disablement calculation algorithm
Each OS-specific implementation of the Image class has its own
implementation of an algorithm to calculate a disabled representation of
a given image. In addition, the algorithm used on Windows and MacOS
produces bad results, which is why usually pre-generated disabled icons
are used.

This change centralizes the implementations of those algorithms in
ImageColorTransformers. The implementations used for the pre-generated
disabled icons is added as default algorithm with an enhanced and the
existing GTK-conforming algorithms being provided as options.

Co-authored-by: Manuel Killinger <Manuel.Killinger@vector.com>